# apps/sync-api/openapi.yaml
openapi: 3.1.0
info:
  title: FieldEdge Sync API
  version: 0.1.0
  description: "Orchestrator between FieldEdge mobile devices and the central Qdrant
cluster. Authentication uses HS256 JWTs issued by `/auth/login` and
refreshed via `/auth/refresh`. All `/sync/*` endpoints require an
`Authorization: Bearer <access_token>` header carrying a token whose `type` claim is `access` (15-minute TTL). See docs/08-AUTH.md for the planned OIDC swap.
"
  contact:
    name: FieldEdge Hackathon Team
    url: https://github.com/choksi2212/field-edge
  license:
    name: Proprietary
servers:
- url: http://localhost:8000
  description: Local development
- url: https://field-edge-sync-api.onrender.com
  description: Production (Render)
...
paths:
  /auth/login:
    post:
      tags:
      - auth
      summary: Exchange device credentials for a JWT pair
      description: 'Mints a short-lived access token (15-minute TTL) and a
        long-lived refresh token (7-day TTL). Both are HS256-signed JWTs
        whose `sub` claim is the supplied `device_id`. Today the only
        requirement is a non-empty `device_id` and an 8+-character
        `device_token` -- see docs/08-AUTH.md for the planned OIDC swap.

        '
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token pair minted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          description: Missing or short device_token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags:
      - auth
      summary: Trade a refresh token for a new pair
      description: 'Rotates the refresh token (new `jti`) and issues a fresh
        access token. v0 policy: the OLD refresh token keeps working
        until its natural expiry (no server-side jti revocation list
        yet -- TODO in OIDC pass).

        '
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags:
      - auth
      summary: Echo the authenticated device and access-token expiry
      description: 'A lightweight bearer probe -- clients can hit this on app
        foreground to decide whether to silently refresh.

        '
      operationId: me
      security:
      - DeviceBearer: []
      responses:
        '200':
          description: Authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    DeviceBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'HS256-signed JWT issued by /auth/login. The token

        carries the device_id as its `sub` claim and the literal

        string `access` as its `type` claim. Tokens expire after

        JWT_ACCESS_TTL_SECONDS (default 900s = 15 min); refresh via

        /auth/refresh. The v0 login endpoint accepts any 8+ char

        device_token -- OIDC swap replaces that with a real

        auth-code exchange.'
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request payload failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: true
      description: Generic FastAPI HTTP error envelope.
      properties:
        detail:
          type: string
    HTTPValidationError:
      type: object
      additionalProperties: false
      required:
      - detail
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorItem'
    ValidationErrorItem:
      type: object
      additionalProperties: true
      properties:
        loc:
          type: array
          items:
            oneOf:
            - type: string
            - type: integer
        msg:
          type: string
        type:
          type: string
    LoginRequest:
      type: object
      additionalProperties: false
      required:
      - device_id
      - device_token
      properties:
        device_id:
          type: string
          description: Per-install device id (ULID).
        device_token:
          type: string
          minLength: 8
          description: 'Today: any 8+ char shared secret. OIDC pass: an auth-code.'
    RefreshRequest:
      type: object
      additionalProperties: false
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
          description: A refresh token issued by /auth/login or a prior /auth/refresh.
    TokenPair:
      type: object
      additionalProperties: false
      required:
      - access_token
      - refresh_token
      - expires_in
      properties:
        access_token:
          type: string
          description: HS256 JWT, 15-minute TTL, type=access.
        refresh_token:
          type: string
          description: HS256 JWT, 7-day TTL, type=refresh, has jti.
        expires_in:
          type: integer
          description: Access-token TTL in seconds.
    MeResponse:
      type: object
      additionalProperties: false
      required:
      - device_id
      - exp
      properties:
        device_id:
          type: string
          description: Sub claim of the access token used for this call.
        exp:
          type: integer
          description: Unix-seconds expiry of the access token used for this call.
...